[hootl] Move all files related to github webhook preset into front/lib/triggers/built-in-webhooks/github/
