+++ front/lib/triggers/built-in-webhooks/test/test_webhook_source_presets.ts
@@ -1,10 +1,10 @@
 import { RocketIcon } from "@dust-tt/sparkle";
 
 import type { Authenticator } from "@app/lib/auth";
-import type { RemoteWebhookService } from "@app/lib/triggers/services/remote_webhook_service";
 import logger from "@app/logger/logger";
 import type { Result } from "@app/types";
 import { Ok } from "@app/types";
+import type { RemoteWebhookService } from "@app/types/triggers/remote_webhook_service";
 import type {
   PresetWebhook,
   WebhookEvent,
